pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "weight-room"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.109",
  "uvicorn[standard]>=0.27",
  "pydantic>=2.5",
  "pydantic-settings>=2.0",
  "PyJWT>=2.8",
  "cryptography>=41.0",
  "supabase>=2.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
